Simplify validation dir structure
