--- packages/lib/src/validation/base/index.ts
@@ -1,3 +1,0 @@
-export * from './choices';
-export * from './commands';
-export * from './options';
