+++ packages/lib/src/validation/choices.ts
