--- packages/lib/src/validation/commands/index.ts
@@ -1,4 +1,0 @@
-export * from './user';
-export * from './message';
-export * from './slash';
-export * from './subcommand';
