+++ packages/lib/src/validation/index.ts
@@ -1,4 +1,4 @@
-export * from './base';
+export * from './checks';
+export * from './choices';
 export * from './commands';
 export * from './options';
-export * from './checks';
