--- packages/lib/src/validation/options/index.ts
@@ -1,1 +1,0 @@
-export * from './commands';
